com.unity.live-capture@2.0.0-pre.3 ## [2.0.0-pre.3] - 2021-10-26
### Added
  - Timecode synchronization
  - The client can select the Take to play, and edit its metadata.

### Changed
  - TakeRecorder improvements to better handle device status.
  - Instead of using AnimationJobs, the LiveCaptureDevice now uses LiveUpdate to set the actor's properties.
  - Editor UI improvements

### Removed
  -  LiveCaptureDevice IsLive/SetLive API.

### Fix
  - Creating actors in Prefab mode doesn't work.
  - Bug where the camera actor can be left in an indeterminate state if lens postprocessor is reset.

+++ Runtime/ARKitFaceCapture/FaceBlendShapeCurves.cs
@@ -52,13 +52,13 @@ namespace Unity.LiveCapture.ARKitFaceCapture
         }
 
         /// <inheritdoc/>
-        public void AddKey(float time, FaceBlendShapePose value)
+        public void AddKey(double time, FaceBlendShapePose value)
         {
             AddKey(time, ref value);
         }
 
-        /// <inheritdoc cref="AddKey(float,FaceBlendShapePose)"/>
-        public void AddKey(float time, ref FaceBlendShapePose value)
+        /// <inheritdoc cref="AddKey(double,Unity.LiveCapture.ARKitFaceCapture.FaceBlendShapePose)"/>
+        public void AddKey(double time, ref FaceBlendShapePose value)
         {
             for (var i = 0; i < FaceBlendShapePose.ShapeCount; ++i)
             {
