+++ Runtime/Core/Curves/Vector3Curve.cs
@@ -54,7 +54,7 @@ namespace Unity.LiveCapture
         }
 
         /// <inheritdoc/>
-        public void AddKey(float time, Vector3 value)
+        public void AddKey(double time, Vector3 value)
         {
             for (var i = 0; i < m_Curves.Length; ++i)
             {
